Extract PointFreePrelude and GitHub
