+++ .sourcery-templates/DerivePartialIsos.stencil
@@ -1,5 +1,6 @@
 import ApplicativeRouter
 import Either
+import PointFreePrelude
 import Prelude
 
 {% for enum in types.implementing.DerivePartialIsos|enum %}
