+++ Sources/GitHub/GitHub.swift
@@ -1,6 +1,7 @@
 import Either
 import Foundation
 import Optics
+import PointFreePrelude
 import Prelude
 
 public struct GitHub {
@@ -8,16 +9,20 @@ public struct GitHub {
   public var fetchAuthToken: (String) -> EitherIO<Error, Either<OAuthError, AccessToken>>
 
   /// Fetches a GitHub user's emails.
-  public var fetchEmails: (AccessToken) -> EitherIO<Error, [GitHub.User.Email]>
+  public var fetchEmails: (AccessToken) -> EitherIO<Error, [User.Email]>
 
   /// Fetches a GitHub user from an access token.
   public var fetchUser: (AccessToken) -> EitherIO<Error, User>
+}
 
-  static let live = GitHub(
-    fetchAuthToken: PointFree.fetchAuthToken >>> runGitHub,
-    fetchEmails: PointFree.fetchEmails >>> runGitHub,
-    fetchUser: PointFree.fetchUser >>> runGitHub
-  )
+extension GitHub {
+  public init(clientId: String, clientSecret: String) {
+    self.init(
+      fetchAuthToken: fetchGitHubAuthToken(clientId: clientId, clientSecret: clientSecret) >>> runGitHub,
+      fetchEmails: fetchGitHubEmails >>> runGitHub,
+      fetchUser: fetchGitHubUser >>> runGitHub
+    )
+  }
 
   public struct AccessToken: Codable {
     public private(set) var accessToken: String
@@ -59,40 +64,48 @@ public struct GitHub {
   public struct UserEnvelope: Codable {
     public private(set) var accessToken: AccessToken
     public private(set) var gitHubUser: User
+
+    public init(accessToken: AccessToken, gitHubUser: User) {
+      self.accessToken = accessToken
+      self.gitHubUser = gitHubUser
+    }
   }
 }
 
-func fetchAuthToken(with code: String) -> DecodableRequest<Either<GitHub.OAuthError, GitHub.AccessToken>> {
+func fetchGitHubAuthToken(
+  clientId: String, clientSecret: String
+  )
+  -> (String)
+  -> DecodableRequest<Either<GitHub.OAuthError, GitHub.AccessToken>> {
 
-  var request = URLRequest(url: URL(string: "https://github.com/login/oauth/access_token")!)
-  request.httpMethod = "POST"
-  request.httpBody = try? gitHubJsonEncoder.encode(
-    [
-      "client_id": Current.envVars.gitHub.clientId,
-      "client_secret": Current.envVars.gitHub.clientSecret,
-      "code": code,
-      "accept": "json"
-    ])
-  request.allHTTPHeaderFields = [
-    "Content-Type": "application/json",
-    "Accept": "application/json"
-  ]
+    return { code in
+      var request = URLRequest(url: URL(string: "https://github.com/login/oauth/access_token")!)
+      request.httpMethod = "POST"
+      request.httpBody = try? gitHubJsonEncoder.encode(
+        [
+          "client_id": clientId,
+          "client_secret": clientSecret,
+          "code": code,
+          "accept": "json"
+        ])
+      request.allHTTPHeaderFields = [
+        "Content-Type": "application/json",
+        "Accept": "application/json"
+      ]
 
-  return DecodableRequest(rawValue: request)
+      return DecodableRequest(rawValue: request)
+    }
 }
 
-func fetchEmails(token: GitHub.AccessToken) -> DecodableRequest<[GitHub.User.Email]> {
-
+func fetchGitHubEmails(token: GitHub.AccessToken) -> DecodableRequest<[GitHub.User.Email]> {
   return apiDataTask("user/emails", token: token)
 }
 
-func fetchUser(with token: GitHub.AccessToken) -> DecodableRequest<GitHub.User> {
-
+internal func fetchGitHubUser(with token: GitHub.AccessToken) -> DecodableRequest<GitHub.User> {
   return apiDataTask("user", token: token)
 }
 
 private func apiDataTask<A>(_ path: String, token: GitHub.AccessToken) -> DecodableRequest<A> {
-
   return DecodableRequest(
     rawValue: URLRequest(url: URL(string: "https://api.github.com/" + path)!)
       |> \.allHTTPHeaderFields .~ [
@@ -103,7 +116,6 @@ private func apiDataTask<A>(_ path: String, token: GitHub.AccessToken) -> Decoda
 }
 
 private func runGitHub<A>(_ gitHubRequest: DecodableRequest<A>) -> EitherIO<Error, A> {
-
   return jsonDataTask(with: gitHubRequest.rawValue, decoder: gitHubJsonDecoder)
 }
