+++ Sources/PointFree/Auth.swift
@@ -1,11 +1,13 @@
 import Css
 import Either
 import Foundation
+import GitHub
 import Html
 import HtmlCssSupport
 import HttpPipeline
 import HttpPipelineHtmlSupport
 import Optics
+import PointFreePrelude
 import Prelude
 import Styleguide
 import UrlFormEncoding
