+++ Sources/PointFree/Routers/PointFreeRoutes.swift
@@ -3,6 +3,7 @@ import Foundation
 import Either
 import HttpPipeline
 import Optics
+import PointFreePrelude
 import Prelude
 import UrlFormEncoding
 
@@ -199,7 +200,7 @@ private let routers: [Router<Route>] = [
     <¢> post %> lit("invites") %> formField("email", Optional.iso.some >>> opt(.rawRepresentable)) <% end,
 
   .invite <<< .show
-    <¢> get %> lit("invites") %> pathParam(.rawRepresentable >>> .rawRepresentable) <% end,
+    <¢> get %> lit("invites") %> pathParam(.uuid >>> .rawRepresentable) <% end,
 
   .login
     <¢> get %> lit("login") %> queryParam("redirect", opt(.string)) <% end,
