+++ Sources/PointFree/TODO.swift
@@ -47,268 +47,6 @@ public func requireSome<A>(
     }
 }
 
-// TODO: Move to PreludeFoundation?
-
-private let sessionConfig = URLSessionConfiguration.default
-  |> \.timeoutIntervalForRequest .~ 25
-  |> \.timeoutIntervalForResource .~ 25
-
-public func dataTask(with request: URLRequest) -> EitherIO<Error, (Data, URLResponse)> {
-  return .init(
-    run: .init { callback in
-
-      let startTime = Current.date().timeIntervalSince1970
-      let uuid = UUID().uuidString
-      Current.logger.debug("[Data Task] \(uuid) \(request.httpMethod ?? "UNKNOWN")")
-
-      let session = URLSession(configuration: sessionConfig)
-      session
-        .dataTask(with: request) { data, response, error in
-          defer { session.finishTasksAndInvalidate() }
-
-          let endTime = Current.date().timeIntervalSince1970
-          let delta = Int((endTime - startTime) * 1000)
-
-          let dataMsg = data.map { _ in "some" } ?? "none"
-          let responseMsg = response.map { _ in "some" } ?? "none"
-          let errorMsg = error.map(String.init(describing:)) ?? "none"
-
-          Current.logger.debug("""
-            [Data Task] \(uuid) \(delta)ms, \
-            \(request.url?.absoluteString ?? "nil"), \
-            (data, response, error) = \
-            (\(dataMsg), \(responseMsg), \(errorMsg))
-            """
-          )
-
-          if let error = error {
-            callback(.left(error))
-            return
-          }
-          if let data = data, let response = response {
-            callback(.right((data, response)))
-            return
-          }
-
-          // If the code gets here, then something really bad has happened and let's notify admins.
-          sendEmail(
-            to: adminEmails,
-            subject: "[PointFree Error] JSON Data Task Never Invoked Callback",
-            content: inj1("""
-              Request: \(request.debugDescription)
-              Data: \(dataMsg)
-              Response: \(responseMsg)
-              Error: \(errorMsg)
-              """)
-            ).run.parallel.run { _ in }
-        }
-        .resume()
-    }
-  )
-}
-
-enum JSONError: Error {
-  case error(String, Error)
-}
-
-public typealias DecodableRequest<A> = Tagged<A, URLRequest> where A: Decodable
-
-public func jsonDataTask<A>(with request: URLRequest, decoder: JSONDecoder? = nil)
-  -> EitherIO<Error, A>
-  where A: Decodable {
-
-    return dataTask(with: request)
-      .map(first)
-      .flatMap { data in
-        .wrap {
-          do {
-            return try (decoder ?? defaultDecoder).decode(A.self, from: data)
-          } catch {
-            throw JSONError.error(String(decoding: data, as: UTF8.self), error)
-          }
-        }
-    }
-}
-
-private let defaultDecoder = JSONDecoder()
-
-extension Either: Monoid where R: Monoid {
-  public static var empty: Either {
-    return pure(R.empty)
-  }
-}
-
-public func retry<E, A>(maxRetries: Int) -> (EitherIO<E, A>) -> EitherIO<E, A> {
-  return { $0.retry(maxRetries: maxRetries) }
-}
-
-public func retry<E, A>(maxRetries: Int, backoff: @escaping (Int) -> DispatchTimeInterval)
-  -> (EitherIO<E, A>)
-  -> EitherIO<E, A> {
-  return { $0.retry(maxRetries: maxRetries, backoff: backoff) }
-}
-
-public func delay<E, A>(_ interval: DispatchTimeInterval) -> (EitherIO<E, A>) -> (EitherIO<E, A>) {
-  return { $0.delay(interval) }
-}
-
-public func delay<E, A>(_ interval: TimeInterval) -> (EitherIO<E, A>) -> (EitherIO<E, A>) {
-  return { $0.delay(interval) }
-}
-
-public func zip2<A, B>(_ lhs: Parallel<A>, _ rhs: Parallel<B>) -> Parallel<(A, B)> {
-  return tuple <¢> lhs <*> rhs
-}
-
-public func zip3<A, B, C>(_ a: Parallel<A>, _ b: Parallel<B>, _ c: Parallel<C>) -> Parallel<(A, B, C)> {
-  return tuple3 <¢> a <*> b <*> c
-}
-
-public func zip4<A, B, C, D>(
-  _ a: Parallel<A>,
-  _ b: Parallel<B>,
-  _ c: Parallel<C>,
-  _ d: Parallel<D>
-  ) -> Parallel<(A, B, C, D)> {
-
-  return tuple4 <¢> a <*> b <*> c <*> d
-}
-
-public func zip5<A, B, C, D, E>(
-  _ a: Parallel<A>,
-  _ b: Parallel<B>,
-  _ c: Parallel<C>,
-  _ d: Parallel<D>,
-  _ e: Parallel<E>
-  ) -> Parallel<(A, B, C, D, E)> {
-
-  return tuple5 <¢> a <*> b <*> c <*> d <*> e
-}
-
-public func zip6<A, B, C, D, E, F>(
-  _ a: Parallel<A>,
-  _ b: Parallel<B>,
-  _ c: Parallel<C>,
-  _ d: Parallel<D>,
-  _ e: Parallel<E>,
-  _ f: Parallel<F>
-  ) -> Parallel<(A, B, C, D, E, F)> {
-
-  return tuple6 <¢> a <*> b <*> c <*> d <*> e <*> f
-}
-
-public func zip7<A, B, C, D, E, F, G>(
-  _ a: Parallel<A>,
-  _ b: Parallel<B>,
-  _ c: Parallel<C>,
-  _ d: Parallel<D>,
-  _ e: Parallel<E>,
-  _ f: Parallel<F>,
-  _ g: Parallel<G>
-  ) -> Parallel<(A, B, C, D, E, F, G)> {
-
-  return tuple7 <¢> a <*> b <*> c <*> d <*> e <*> f <*> g
-}
-
-public func zip8<A, B, C, D, E, F, G, H>(
-  _ a: Parallel<A>,
-  _ b: Parallel<B>,
-  _ c: Parallel<C>,
-  _ d: Parallel<D>,
-  _ e: Parallel<E>,
-  _ f: Parallel<F>,
-  _ g: Parallel<G>,
-  _ h: Parallel<H>
-  ) -> Parallel<(A, B, C, D, E, F, G, H)> {
-
-  return tuple8 <¢> a <*> b <*> c <*> d <*> e <*> f <*> g <*> h
-}
-
-public func tuple5<A, B, C, D, E>(_ a: A) -> (B) -> (C) -> (D) -> (E) -> (A, B, C, D, E) {
-  return { b in { c in { d in { e in (a, b, c, d, e) } } } }
-}
-
-public func tuple6<A, B, C, D, E, F>(_ a: A) -> (B) -> (C) -> (D) -> (E) -> (F) -> (A, B, C, D, E, F) {
-  return { b in { c in { d in { e in { f in (a, b, c, d, e, f) } } } } }
-}
-
-public func tuple7<A, B, C, D, E, F, G>(_ a: A) -> (B) -> (C) -> (D) -> (E) -> (F) -> (G) -> (A, B, C, D, E, F, G) {
-  return { b in { c in { d in { e in { f in { g in (a, b, c, d, e, f, g) } } } } } }
-}
-
-public func tuple8<A, B, C, D, E, F, G, H>(_ a: A) -> (B) -> (C) -> (D) -> (E) -> (F) -> (G) -> (H) -> (A, B, C, D, E, F, G, H) {
-  return { b in { c in { d in { e in { f in { g in { h in (a, b, c, d, e, f, g, h) } } } } } } }
-}
-
-public typealias T8<A, B, C, D, E, F, G, Z> = Tuple<A, T7<B, C, D, E, F, G, Z>>
-public typealias Tuple7<A, B, C, D, E, F, G> = T8<A, B, C, D, E, F, G, Prelude.Unit>
-public func get7<A, B, C, D, E, F, G, Z>(_ t: T8<A, B, C, D, E, F, G, Z>) -> G {
-  return t.second.second.second.second.second.second.first
-}
-public func lower<A, B, C, D, E, F, G>(_ tuple: Tuple7<A, B, C, D, E, F, G>) -> (A, B, C, D, E, F, G) {
-  return (get1(tuple), get2(tuple), get3(tuple), get4(tuple), get5(tuple), get6(tuple), get7(tuple))
-}
-
-// better way of doing this? or should we add to Either.swift?
-public func requireSome<A>(_ e: Either<Error, A?>) -> Either<Error, A> {
-  switch e {
-  case let .left(e):
-    return .left(e)
-  case let .right(a):
-    return a.map(Either.right) ?? .left(unit)
-  }
-}
-
-// TODO: Move to swift-web
-extension URLRequest {
-  public var cookies: [String: String] {
-    let pairs = (self.allHTTPHeaderFields?["Cookie"] ?? "")
-      .components(separatedBy: "; ")
-      .map {
-        $0.split(separator: "=", maxSplits: 1, omittingEmptySubsequences: false)
-          .map(String.init)
-      }
-      .compactMap { (pair: [String]) -> (String, String) in
-        (pair[0], pair.count == 2 ? pair[1] : "")
-    }
-    return .init(pairs, uniquingKeysWith: { $1 })
-  }
-}
-
-extension UUID: RawRepresentable {
-  public var rawValue: String {
-    return self.uuidString
-  }
-
-  public init?(rawValue: String) {
-    guard let uuid = UUID(uuidString: rawValue) else { return nil }
-    self = uuid
-  }
-}
-
-extension DispatchTimeInterval {
-  private var nanoseconds: Int? {
-    switch self {
-    case let .seconds(n):
-      return .some(n * 1_000_000_000)
-    case let .milliseconds(n):
-      return .some(n * 1_000_000)
-    case let .microseconds(n):
-      return .some(n * 1_000)
-    case let .nanoseconds(n):
-      return .some(n)
-    case .never:
-      return nil
-    }
-  }
-
-  public static func + (lhs: DispatchTimeInterval, rhs: DispatchTimeInterval) -> DispatchTimeInterval {
-    return (curry(+) <¢> lhs.nanoseconds <*> rhs.nanoseconds)
-      .map(DispatchTimeInterval.nanoseconds)
-      ?? .never
-  }
-}
-
 extension PartialIso where A == String, B == String {
   public static func decrypted(withSecret secret: String) -> PartialIso<String, String> {
     return PartialIso(
@@ -319,27 +57,6 @@ extension PartialIso where A == String, B == String {
 
   public static var appDecrypted: PartialIso<String, String> {
     return .decrypted(withSecret: Current.envVars.appSecret)
-  }
-}
-
-public func sequence1<A, Z>(_ t: T2<IO<A>, Z>) -> IO<T2<A, Z>> {
-  return IO {
-    return t |> over1(perform)
-  }
-}
-public func sequence2<A, B, Z>(_ t: T3<A, IO<B>, Z>) -> IO<T3<A, B, Z>> {
-  return IO {
-    return t |> over2(perform)
-  }
-}
-public func sequence3<A, B, C, Z>(_ t: T4<A, B, IO<C>, Z>) -> IO<T4<A, B, C, Z>> {
-  return IO {
-    return t |> over3(perform)
-  }
-}
-public func sequence4<A, B, C, D, Z>(_ t: T5<A, B, C, IO<D>, Z>) -> IO<T5<A, B, C, D, Z>> {
-  return IO {
-    return t |> over4(perform)
   }
 }
 
@@ -367,81 +84,8 @@ public func payload<A, B>(
     })
 }
 
-public func zurry<A>(_ f: () -> A) -> A {
-  return f()
-}
-
-public func unzurry<A>(_ a: A) -> () -> A {
-  return { a }
-}
-
-public func require4<A, B, C, D, Z>(_ x: T5<A, B, C, D?, Z>) -> T5<A, B, C, D, Z>? {
-  return get4(x).map { over4(const($0)) <| x }
-}
-public func require5<A, B, C, D, E, Z>(_ x: T6<A, B, C, D, E?, Z>) -> T6<A, B, C, D, E, Z>? {
-  return get5(x).map { over5(const($0)) <| x }
-}
 
 // PreludeFoundation
-
-private let guaranteeHeaders = \URLRequest.allHTTPHeaderFields %~ {
-  $0 ?? [:]
-}
-
-let setHeader = { name, value in
-  guaranteeHeaders
-    <> (\.allHTTPHeaderFields <<< map <<< \.[name] .~ value)
-}
-
-func attachBasicAuth(username: String = "", password: String = "") -> (URLRequest) -> URLRequest {
-  let encoded = Data((username + ":" + password).utf8).base64EncodedString()
-  return setHeader("Authorization", "Basic " + encoded)
-}
-
-let attachFormData =
-  urlFormEncode(value:)
-    >>> ^\.utf8
-    >>> Data.init(_:)
-    >>> set(\URLRequest.httpBody)
-
-// Prelude
-
-public func concat<A>(_ fs: [(A) -> A]) -> (A) -> A {
-  return { a in
-    fs.reduce(a) { a, f in f(a) }
-  }
-}
-
-public func concat<A>(_ fs: ((A) -> A)..., and fz: @escaping (A) -> A = id) -> (A) -> A {
-  return concat(fs + [fz])
-}
-
-public func concat<A>(_ fs: [(inout A) -> Void]) -> (inout A) -> Void {
-  return { a in
-    fs.forEach { f in f(&a) }
-  }
-}
-
-public func concat<A>(_ fs: ((inout A) -> Void)..., and fz: @escaping (inout A) -> Void = { _ in })
-  -> (inout A) -> Void {
-
-    return concat(fs + [fz])
-}
-
-// Prelude / Overture
-
-public func update<A>(_ value: inout A, _ changes: ((A) -> A)...) {
-  value = value |> concat(changes)
-}
-
-public func update<A>(_ value: inout A, _ changes: ((inout A) -> Void)...) {
-  concat(changes)(&value)
-}
-
-func zip2<A, B>(_ a: A?, _ b: B?) -> (A, B)? {
-  guard let a = a, let b = b else { return nil }
-  return (a, b)
-}
 
 public func responseTimeout(_ interval: TimeInterval)
   -> (@escaping Middleware<StatusLineOpen, ResponseEnded, Prelude.Unit, Data>)
