+++ Sources/PointFreePrelude/EmailAddress.swift
@@ -1,0 +1,3 @@
+import Tagged
+
+public typealias EmailAddress = Tagged<(email: (), String), String>
