+++ Sources/PointFreePrelude/Zip.swift
@@ -1,0 +1,4 @@
+public func zip<A, B>(_ a: A?, _ b: B?) -> (A, B)? {
+  guard let a = a, let b = b else { return nil }
+  return (a, b)
+}
