+++ Sources/PointFreePrelude/Zurry.swift
@@ -1,0 +1,7 @@
+public func zurry<A>(_ f: () -> A) -> A {
+  return f()
+}
+
+public func unzurry<A>(_ a: A) -> () -> A {
+  return { a }
+}
